Fixed a bug with sorting and improved getting the data type of a cell.
Just because a row was previously sorted, doesn’t mean that reversing
it will yield to the alternate sorting (ascending vs descending). If
all the cells have the same value, they don’t need to be resorted. Also
improved the getDataType() function to get the inner most child. I
changed the quick sort function to allow sorting ascending and
descending and added in a check for equal elements. I had to use less
than or greater than instead of equals because JSLint was forcing me to
use ‘===‘, but that doesn’t work with strings.